Use double instead of int in kernel

+++ resources/add.cu
@@ -1,11 +1,13 @@
-__global__ void add_kernel(int *a, int *b, int *c, int n) {
+#include <stdio.h>
+
+__global__ void add_kernel(double *a, double *b, double *c, int n) {
   int i = blockIdx.x * blockDim.x + threadIdx.x;
   if (i < n) {
     c[i] = a[i] + b[i];
   }
 }
 
-extern "C" void add(int *a, int *b, int *c, int n) {
+extern "C" void add(double *a, double *b, double *c, int n) {
   dim3 blockDim(256);
   dim3 gridDim((n + blockDim.x - 1) / blockDim.x);
